move `IDeviceKey` and `IOneTimeKey` to `@types/crypto`
